REAL HUMAN DELAY, typing idicator, Invite Link
Added a human delay to the bot, so it takes a bit longer to respond like its a human typing.

See the human_delay function under functions.js.

To implement, just write: functions.human_delay();

I have added typing indicators for the bot responding to commands.

Added a invite link to the README.md so anyone can invite the bot.